[androsdk-788] Move packages
